Extract shared distribution context helper into _distribution_context
Moves chart types (VersionInfo, BucketData, RecipeDistributionContext), colour
constants, and the three private builder helpers into a new private module
_distribution_context. Adds build_distribution_context as the shared entry
point used by both the existing get_recipe_distribution use case and the
upcoming get_move_preview_distribution. get_recipe_distribution becomes a thin
wrapper; all public symbols are re-exported so existing importers are unaffected.
